Extract Instructions from Assistant Builder
